#329: Move `XYZValues` to a more common place.
