+++ hydrolib/core/io/common/models.py
@@ -1,3 +1,4 @@
+from typing import List, Optional
 from enum import Enum
 
 
@@ -15,3 +16,15 @@ class LocationType(str, Enum):
 
     all = "all"
     """str: Denotes that both 1D and 2D locations may be selected."""
+
+class XYZValues:
+    """Class that holds lists of x, y and z values."""
+
+    x: List[float] = []
+    """The x values."""
+
+    y: List[float] = []
+    """The y values."""
+
+    z: List[Optional[float]] = []
+    """The z values. The values can be `None`."""
